pyproject.toml
[project]
name = "adk-samples"
version = "0.1.0"
description = "ADK Samples"
authors = [         
    { name = "Dan Higgins", email = "dphiggins@gmail.com" },
]
license = "Apache License 2.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "google-adk>=1.11.0",
    "chromadb>=0.4.21",
    "litellm>=1.80.10",
    "mcp[cli]>=1.13.0",
    "pypdf>=6.4.2",
    "sentence-transformers>=5.2.0",
    "langchain-text-splitters>=1.1.0",
    "openinference-instrumentation>=0.1.42",
    "arize-phoenix>=12.25.1",
    "openinference-instrumentation-google-adk>=0.1.8",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "ddgs>=9.10.0",
]

